pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "latex-formula-tool"
version = "0.1.0"
description = "PyQt desktop tool that converts screenshots or text descriptions into LaTeX formula code with preview rendering."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyQt6>=6.6",
    "PyQt6-WebEngine>=6.6",
    "Pillow>=10.0",
    "openai>=1.40",
    "python-dotenv>=1.0",
    "markdown>=3.6",
]

[project.scripts]
latex-formula-tool = "latex_formula_tool.main:main"

[tool.setuptools.packages.find]
include = ["latex_formula_tool*"]

[tool.ruff]
line-length = 100
